Fix bugs that occured in move to newest giraph git version
Also change InputFormat and make bitfield length changeable.

+++ src/main/java/org/apache/giraph/examples/closeness/OpenLongIntHashMapWritable.java
@@ -31,7 +31,7 @@ public class OpenLongIntHashMapWritable extends OpenLongIntHashMap
     state = new byte[capacity];
     
     for (int i = 0; i < table.length; ++i) {
-      table[i] = in.readInt();
+      table[i] = in.readLong();
       values[i] = in.readInt();
       state[i] = in.readByte();
     }
